chore: move child workflow and signal-external errors to cadence.error
