+++ cadence/_internal/workflow/statemachine/signal_external_workflow_state_machine.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from typing import Any
-
 from cadence._internal.workflow.statemachine.decision_state_machine import (
     BaseDecisionStateMachine,
     DecisionFuture,
@@ -11,14 +9,9 @@ from cadence._internal.workflow.statemachine.decision_state_machine import (
 )
 from cadence._internal.workflow.statemachine.event_dispatcher import EventDispatcher
 from cadence.api.v1 import decision, history
+from cadence.error import SignalExternalWorkflowFailed
 
 signal_external_events = EventDispatcher("initiated_event_id")
-
-
-class SignalExternalWorkflowFailed(Exception):
-    def __init__(self, message: str, cause: Any) -> None:
-        super().__init__(message)
-        self.cause = cause
 
 
 class SignalExternalWorkflowStateMachine(BaseDecisionStateMachine):
